+++ R/math.R
@@ -12,7 +12,6 @@
 #'
 #' @examples
 #' x <- set_quantities(1:3, m/s, 0.1)
-#' exp(x)
 #' log(x)
 #' cumsum(x)
 #' cumprod(x)