+++ R/misc.R
@@ -164,7 +164,7 @@ cbind.quantities <- function(..., deparse.level = 1) {
   dots <- list(...)
   stopifnot(all(sapply(dots, inherits, "units")))
   u <- units(dots[[1]])
-  dots <- getS3method("set_units", "mixed_units")(dots, as.character(u))
+  dots <- lapply(dots, set_units, u, mode="standard")
 
   nm <- names(as.list(match.call()))
   nm <- nm[nm != "" & nm != "deparse.level"]