+++ R/quantities.R
@@ -101,7 +101,7 @@ quantities.quantities <- function(x) {
 
 #' @name quantities
 #' @param unit a \code{units} object, or something coercible to one with
-#' \code{as_units} (see \code{\link[units]{set_units}}).
+#' \code{as_units} (see \code{\link[units:units]{set_units}}).
 #' @param errors a numeric vector of length 1 or the same length as \code{x}
 #' (see \code{\link[errors:errors]{set_errors}}).
 #' @inheritParams units::set_units